Use normalized floats to represent joystick axis values
git-svn-id: https://s.snth.net/svn/neverball/trunk@3351 78b8d119-cf0a-0410-b17c-f493084dd1d7

+++ ball/game_server.h
@@ -21,8 +21,8 @@ void  game_clr_goal(void);
 
 void  game_set_ang(int, int);
 void  game_set_pos(int, int);
-void  game_set_x  (int);
-void  game_set_z  (int);
+void  game_set_x  (float);
+void  game_set_z  (float);
 void  game_set_cam(int);
 void  game_set_rot(float);
